fixed a lot of warnings and errors

+++ YapZap/YapZap/UserInterface/MyRecordingsCell.m
@@ -37,13 +37,13 @@
     self.tagLabel.text = recording.tagName;
     
     
-    self.commentLabel.text = [NSString stringWithFormat:@"%ld", recording.childrenLength];
+    self.commentLabel.text = [NSString stringWithFormat:@"%ld", (long)recording.childrenLength];
     self.commentLabel.textColor = recording.childrenLength?[UIColor blackColor]:[UIColor whiteColor];
     UIImage* commentImage = recording.childrenLength?[UIImage imageNamed:@"comments_full.png"]:[UIImage imageNamed:@"comments_empty.png"];
     [self.commentButton setImage:commentImage forState:UIControlStateNormal];
     
     
-    self.likesLabel.text = [NSString stringWithFormat:@"%ld", recording.likes];
+    self.likesLabel.text = [NSString stringWithFormat:@"%ld", (long)recording.likes];
     self.likesLabel.textColor = recording.likes?[UIColor blackColor]:[UIColor whiteColor];
     UIImage* likeImage = recording.likes?[UIImage imageNamed:@"heart_full.png"]:[UIImage imageNamed:@"heart_empty.png"];
     [self.likeButton setImage:likeImage forState:UIControlStateNormal];
